+++ test/integration/run.sh
@@ -70,6 +70,7 @@ TESTS=(
 	"test_cli_init_new.sh"
 	"test_cli_init_config_doctor.sh"
 	"test_cli_config_variants.sh"
+	"test_cli_run_tool_allow_self.sh"
 	"test_cli_validate.sh"
 	"test_project_root_detection.sh"
 	"test_elicitation.sh"